Completed pipeline to run from main.py
