+++ ECommerce_Churn_Propensity_Model/main.py
@@ -1,0 +1,6 @@
+import argparse
+
+parser = argparse.ArgumentParser(description='Churn propensity pipeline')
+parser.add_argument('--module', type=str, choices=['eda', 'preprocessing', 'model_training'])
+args = parser.parse_args()
+
