+++ ECommerce_Churn_Propensity_Model/src/preprocessing/preprocessing.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-
 # Load variables from __init__.py
 from . import pre_processing, Config
